处理消息问题

+++ network/TransitData.go
@@ -1,6 +1,6 @@
 package network
 
 type TransitData struct {
-	MsgId int32
-	Data []byte
+	MsgId int64
+	Data  []byte
 }
